widget: Draw box cursor inside the character cell
When the terminal widget does not have the input focus, the block cursor
is changed to a box around the character.  VTE was drawing that box
completely outside the character cell, which was good with respect to
preserving full visibility for the character under the cursor, but
looked really bad in some cases (e.g., when the line width was larger
than the padding aroung the terminal, some parts of the line could be
clipped by the window frame).

Apparently most other terminal emulators draw the box cursor completely
inside the character cell, which looks more consistent when the box and
the block cursor are compared, but may obscure some parts of the
character under the cursor.  Make VTE behave the same way as other
terminal emulators - draw the outline cursor completely inside the
character cell.

Some code still assumes that some pixels around the text line may be
touched by the box cursor and need to be repainted; however, fixing the
code to avoid extra redraws in this case is probably more trouble than
it's worth.